+++ Modules/mutant/binary_io.h
@@ -29,14 +29,14 @@ namespace mutant
 	{
 	public:
 		virtual ~binary_input() /*_ __gcc = 0*/ {}
-		virtual void read( void* dest, int n, int* wasRead ) = 0; // throws EIoError
+		virtual void read( void* dest, size_t n, int* wasRead ) = 0; // throws EIoError
 	};
 
 	class binary_output
 	{
 	public:
 		virtual ~binary_output() /*_ __gcc = 0*/ {}
-		virtual void write( void const* src, int n, int* wasWritten ) = 0;
+		virtual void write( void const* src, size_t n, int* wasWritten ) = 0;
 	};
 }
